+++ apps/astarte_data_updater_plant/lib/astarte_data_updater_plant/data_pipeline_supervisor.ex
@@ -24,6 +24,7 @@ defmodule Astarte.DataUpdaterPlant.DataPipelineSupervisor do
   alias Astarte.DataUpdaterPlant.RPC.Handler
 
   alias Astarte.RPC.Protocol.DataUpdaterPlant, as: Protocol
+  alias Astarte.DataUpdaterPlant.ConnectionPoolUtils
 
   def start_link(init_arg) do
     Supervisor.start_link(__MODULE__, init_arg, name: __MODULE__)
@@ -34,6 +35,10 @@ defmodule Astarte.DataUpdaterPlant.DataPipelineSupervisor do
     children = [
       {Registry, [keys: :unique, name: Registry.MessageTracker]},
       {Registry, [keys: :unique, name: Registry.DataUpdater]},
+      # we use different connections for producers and consumers
+      {ExRabbitPool.PoolSupervisor,
+       rabbitmq_config: ConnectionPoolUtils.amqp_producer_config(),
+       connection_pools: [ConnectionPoolUtils.pool_config(:producers_pool)]},
       AMQPEventsProducer,
       ConsumersSupervisor,
       {Astarte.RPC.AMQP.Server, [amqp_queue: Protocol.amqp_queue(), handler: Handler]},