+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [astarte_data_updater_plant] Use multiple connections when connecting to RabbitMQ.
 
 ## [1.0.2] - 2022-04-01
 ### Added